+++ HomeProperty/Services/BlogListing/IBlogListingService.cs
@@ -1,0 +1,11 @@
+﻿using HomeProperty.Models.CustomModels;
+
+namespace HomeProperty.Services.BlogListing
+{
+    public interface IBlogListingService
+    {
+        IList<BlogPageWithComments> GetAllBlogPages(ContentReference blogArchive);
+
+        IList<BlogPageWithComments> GetLatestBlogPages(ContentReference blogArchive);
+    }
+}
